Basic Cell-based Eval
Adds commands: `run-cell`, `run-cell-and-move`, `next-cell`, `prev-cell`. When either command is executed, the code in the current cell is eval'd. Results are displayed block style, above the cell delimiter. The range of the cell is from the start of the line below top delimiter (and/or start of top row of file) to the end of the line before the end delimiter (or the end of the file).

Cells can be delimited with any pattern that matches this regex `^(##|#---|#%%|# %%)`

Adds default keyboard shortcuts for all the above commands.

Gotcha: doesn't support anything to do with selections, probably won't do what users expect when the commands added are run when text is selected.