Handle links to new stacktrace shortened-paths introduced in Julia 1.6.0
This fix will correctly handle links to shortened-paths introduced in the new stacktrace of Julia 1.6.0 by expanding the tilde "~" symbol to be the home directory in the link only (will not change the actual path on screen in the stacktrace). This is tested on a Windows 10 machine, but it should work for MacOs and Linux as well (confirmation would be appreciated).